Add files via upload
HTML and CSS applied to creat a SignUp Form. Form background color black with green buttons. Sign up or Login buttons at the top of the form followed by a "Sign Up For Free" centered on the form. "First Name" and "Last Name" inputs appear on the same line. Followed by Email and Set A Password name fields with "GET STARTED" button at the end of the form. I also used a placeholder attribute in the <input> elements. This way, the placeholder text will appear inside the input fields as default text.