Add files via upload
The purpose of project is to apply the knowledge of HTML, CSS (colors, image manipulation, layout,
animation, and modal window)/Bootstrap and JavaScript to create a simple sales web application. In this case its a Blog App. It has a nav bar with logo and Home, Recipes, About Us and Sign In tab. Clicking on Login button  opens the login form in a modal window. Main modal window body had a H1 title and a 3 card layout with images, it also  shows information such as blog title and description, and a clickable ‘read more’ button that takes you to a link.